update
function untuk dino dan wall
