--- model/fosils/irish_deer/license.txt
@@ -1,11 +1,0 @@
-Model Information:
-* title:	Irish deer fossil
-* source:	https://sketchfab.com/3d-models/irish-deer-fossil-33b64def1e7d40048224f589d49d68ec
-* author:	Uninspirational (https://sketchfab.com/Uninspirational)
-
-Model License:
-* license type:	CC-BY-4.0 (http://creativecommons.org/licenses/by/4.0/)
-* requirements:	Author must be credited. Commercial use is allowed.
-
-If you use this 3D model in your project be sure to copy paste this credit wherever you share it:
-This work is based on "Irish deer fossil" (https://sketchfab.com/3d-models/irish-deer-fossil-33b64def1e7d40048224f589d49d68ec) by Uninspirational (https://sketchfab.com/Uninspirational) licensed under CC-BY-4.0 (http://creativecommons.org/licenses/by/4.0/)
